Acquire search in field caps only when needed
FieldCapabilitiesFetcher acquires a searcher to perform basic checks that don't
actually require a searcher to be performed, as they are purely mappings driven.
Instead, we can reuse the existing QueryRewriteContext to perform such checks
and entirely avoid acquiring a searcher.
